Add extra level of packages
See: #4
