+++ src/ss/ita/kataimpl/taras/TarasKataFiveImpl.java
@@ -1,10 +1,10 @@
-package ss.ita.kataimpl;
+package ss.ita.kataimpl.taras;
 
 import ss.ita.kata.KataFive;
 
 import java.math.BigInteger;
 
-public class MykolaKataFiveImpl implements KataFive {
+public class TarasKataFiveImpl implements KataFive {
     @Override
     public int artificialRain(int[] v) {
         return 0;
